Declare makeNotFoundPageObject fuction
