+++ src/404/index.ts
@@ -1,2 +1,3 @@
 export { get404PageFile, get404PagePath } from './get404PageFile';
 export { get404PageInfo } from './get404PageInfo';
+export { makeNotFoundPageObject } from './makeNotFoundPageObject';
