+++ src/getPageObject/getPageObject.ts
@@ -1,25 +1,23 @@
 import getRouteInfo from './getRouteInfo';
-import makeRouteInfo from './makeRouteInfo';
 import { getPageFile } from '../page';
 import { getAppFile } from '../_app';
-import { parseRoute } from '../utils';
+import { InternalError } from '../_error/error';
+import { makeNotFoundPageObject } from '../404';
 import type {
   ExtendedOptions,
   NextPageFile,
   GenericPageObject,
 } from '../commonTypes';
-import { InternalError } from '../_error/error';
-import { get404PageFile, get404PagePath } from '../404';
 
-export default async function getPageObject({
+export async function getPageObject({
   options,
 }: {
   options: ExtendedOptions;
 }): Promise<GenericPageObject> {
   const routeInfo = await getRouteInfo({ options });
-  const appFile = getAppFile({ options });
 
   if (routeInfo) {
+    const appFile = getAppFile({ options });
     const page = getPageFile<NextPageFile>({
       pagePath: routeInfo.pagePath,
       options,
@@ -38,20 +36,5 @@ export default async function getPageObject({
   }
 
   // Make a NotFoundPageObject for 404 page
-  // @NOTE we currently set pagePath as current path name, but it should
-  // be the path of the currently rendered page file
-  const { route } = options;
-  const { pathname } = parseRoute({ route });
-  const notFoundPageRouteInfo = makeRouteInfo({
-    route,
-    pagePath: pathname,
-  });
-
-  return {
-    ...notFoundPageRouteInfo,
-    type: 'notFound',
-    appFile,
-    page: get404PageFile({ options }),
-    __temp__actualPagePath: get404PagePath({ options }),
-  };
+  return makeNotFoundPageObject({ options });
 }
